[mobile][photos] Split ML DB orchestration from SQL behind IMLDataDB
